requirements.txt
# NVIDIA AI Q Toolkit and MCP Dependencies
aiqtoolkit>=1.1.0
fastmcp>=2.10.0
mcp>=1.0.0

# Web Framework and API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# HTTP Client and Async Support
httpx>=0.25.0
aiohttp>=3.9.0

# Environment and Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Logging
structlog>=23.2.0

# JSON and Data Serialization
orjson>=3.9.0

# Time and Date Utilities
python-dateutil>=2.8.0

# System and Process Management
psutil>=5.9.0

# File Monitoring and MIME Detection
inotify_simple>=1.3.5
python-magic>=0.4.27

# Web UI - Anthropic Claude Integration
flask>=3.0.0
anthropic>=0.18.0